stop offsetting coordinates in CAFMaker
historically the coordinates in the CAFs were shifted to put the origin at the middle of the LAr.

this is not really a future-proof solution:
  * the geometry descriptions are still changing sometimes
  * we need to synchronize coordinates with reco on the various detectors, which don't necessarily use these same conventions

here we remove the offset and fix the "collar" margin to the current correct values.  the latter will need a more robust solution going forward, but this stopgap is what we have for the moment