replace fillTruth() with a TruthMatcher class
The TruthMatcher will fill in SRTrueInteractions & SRTrueParticles using GENIE info (if available).  It serves as a central utility ensuring that we don't wind up duplicating truth info from the pass-through truth in multiple reco files.
The idea is that SRTrueInteractions and SRTrueParticles will be filled on-demand as Reco products' truth-matched info is encountered, rather than all up front.