treat the const args more carefully
"const decltype(blah) &" won't always apply `const` to the outgoing type because the deduced type would be trying to apply `const` to the reference itself, which is a no-op.
if we carefully remove the ref, add const, and then add the ref back, as done here now, it always works.